+++ templates/footer.php
@@ -1,4 +1,8 @@
-<hr>
-<div class="footer">
-	Powered by <a href="<?=URL::SourceCode()?>">ArtArchive</a> v<?=ArtArchive::$version?> © 2019-2025 Estecka
-</div>
+<footer>
+	<p>
+		Powered by
+		<a href="<?=URL::SourceCode()?>">ArtArchive</a>
+		v<?=ArtArchive::$version?>
+		© 2019-2025 Estecka
+	</p>
+</footer>